refactor plugin-library sample
