+++ replugin-sample/plugin/plugin-demo1/library/.gitignore
