+++ replugin-sample/plugin/plugin-demo1/library/proguard-rules.pro
