+++ replugin-sample/plugin/plugin-demo1/library/src/main/AndroidManifest.xml
@@ -1,0 +1,13 @@
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+
+    package="com.qihoo360.replugin.sample.library">
+
+    <application android:allowBackup="false" android:label="@string/app_name"
+        android:supportsRtl="true">
+        <activity
+            android:name=".LibMainActivity"
+            android:label="@string/app_name">
+        </activity>
+    </application>
+
+</manifest>
