+++ replugin-sample/plugin/plugin-demo1/library/src/main/java/com/qihoo360/replugin/sample/library/LibMainActivity.java
@@ -1,0 +1,14 @@
+package com.qihoo360.replugin.sample.library;
+
+import android.app.Activity;
+import android.os.Bundle;
+
+public class LibMainActivity extends Activity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.lib_activity_main);
+    }
+
+}
