+++ replugin-sample/plugin/plugin-demo1/library/src/main/res/layout/lib_activity_main.xml
