+++ replugin-sample/plugin/plugin-demo1/library/src/main/res/values/strings.xml
@@ -1,0 +1,3 @@
+<resources>
+    <string name="app_name">library</string>
+</resources>
