+++ replugin-sample/plugin/plugin-demo1/settings.gradle
@@ -14,4 +14,4 @@
  * the License.
  */
 
-include ':app'
+include ':app', ':library'
