--- replugin-sample/plugin/plugin-library-demo1/.gitignore
@@ -1,9 +1,0 @@
-*.iml
-.gradle
-/local.properties
-/.idea/workspace.xml
-/.idea/libraries
-.DS_Store
-/build
-/captures
-.externalNativeBuild
