--- replugin-sample/plugin/plugin-library-demo1/app/src/main/AndroidManifest.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.osan.replugin.sample.library.librarysample">
-
-    <application>
-        <activity
-            android:name=".LibMainActivity"
-            android:label="@string/app_name">
-        </activity>
-    </application>
-
-</manifest>
