--- replugin-sample/plugin/plugin-library-demo1/app/src/main/java/com/osan/replugin/sample/library/librarysample/LibMainActivity.java
@@ -1,14 +1,0 @@
-package com.osan.replugin.sample.library.librarysample;
-
-import android.app.Activity;
-import android.os.Bundle;
-
-public class LibMainActivity extends Activity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.lib_activity_main);
-    }
-
-}
