--- replugin-sample/plugin/plugin-library-demo1/app/src/main/res/layout/content_lib_main.xml
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical"
-    android:padding="16dp">
-
-    <TextView
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:text="Hello AAR:\nStandard"
-        android:textSize="30sp"
-        android:gravity="center"/>
-
-</LinearLayout>
-
