--- replugin-sample/plugin/plugin-library-demo1/app/src/main/res/values/strings.xml
@@ -1,4 +1,0 @@
-<resources>
-    <string name="app_name">LIBRARY_DEMO1</string>
-    <string name="action_settings">Settings</string>
-</resources>
