--- replugin-sample/plugin/plugin-library-demo1/settings.gradle
@@ -1,1 +1,0 @@
-include ':app'
